DRF_API/requirements.txt
asgiref==3.5.2
boto3==1.24.22
botocore==1.27.22
Django==4.0.5
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
jmespath==1.0.1
pep8==1.7.1
Pillow==9.1.1
PyJWT==2.4.0
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2022.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
upload-file-s3==0.2
urllib3==1.26.9
